[*] changed service: Accout, User and AccountProperties

+++ homeWork/src/main/java/org/example/spring_miniBank/properties/AccountProperties.java
@@ -3,15 +3,17 @@ package org.example.spring_miniBank.properties;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class AccountProperties {
 
     private final int defaultAmount;
-    private final double transferCommission;
+    private final BigDecimal transferCommission;
 
     public AccountProperties(
             @Value("${account.default-amount}") int defaultAmount,
-            @Value("${account.transfer-commission}") double transferCommission
+            @Value("${account.transfer-commission}") BigDecimal transferCommission
             )
     {
         this.defaultAmount = defaultAmount;
@@ -22,7 +24,7 @@ public class AccountProperties {
         return defaultAmount;
     }
 
-    public double getTransferCommission() {
+    public BigDecimal getTransferCommission() {
         return transferCommission;
     }
 }
